requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
google-generativeai>=0.3.0 